xfree86: ddc: move remaining struct cea_* to private header
Not needed in public SDK headers.
